remove named export
